requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
confusable-homoglyphs==3.3.1
Django==5.1.2
django-crispy-forms==2.3
django-filter==24.3
django-paginator==0.2a3
django-phonenumber-field==8.0.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
phonenumbers==8.13.48
pillow==11.0.0
propcache==0.2.0
psycopg2==2.9.10
pyactiveresource==1.0.1
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
python-http-client==3.3.7
requests==2.32.3
setuptools==75.2.0
soupsieve==2.6
sqlparse==0.5.1
starkbank-ecdsa==2.2.0
urllib3==2.2.3
yarl==1.16.0
